Extracts theme utils into separate file
